+++ sanity.cpp
@@ -300,7 +300,7 @@ void check_sprite(PseudoSprite&data){
 		if(_autocorrect&&length%3==2&&length/3!=status.seensprites&&length/3<256){
 			IssueMessage(0,CONSOLE_AUTOCORRECT,_spritenum);
 			IssueMessage(0,AUTOCORRECTING,1,"num-sets",data.ExtractByte(1),length/3);
-			data.SetByteAt(1,uchar(status.seensprites=length/3));
+			data.SetByteAt(1,status.seensprites=length/3);
 		}
 		if(!status.seensprites)IssueMessage(WARNING1,NO_SETS,0x0A);
 		if(CheckLength(length,2+3*status.seensprites,BAD_LENGTH,"byte 1","%2x",status.seensprites,2+3*status.seensprites))
@@ -330,7 +330,7 @@ void check_sprite(PseudoSprite&data){
 		if(_autocorrect&&length%4==2&&length/4!=numids&&length/4<256){
 			IssueMessage(0,CONSOLE_AUTOCORRECT,_spritenum);
 			IssueMessage(0,AUTOCORRECTING,1,"num",numids,length/4);
-			data.SetByteAt(1,uchar(numids=length/4));
+			data.SetByteAt(1,numids=length/4);
 		}
 		if(CheckLength(length,2+4*numids,BAD_LENGTH,"<num>","%2x",numids,2+4*numids))return;
 		if(!numids)IssueMessage(WARNING1,NO_GRFIDS);
@@ -379,7 +379,7 @@ void check_sprite(PseudoSprite&data){
 		if(_autocorrect&&length%4==2&&length/4!=status.seensprites&&length/4<256){
 			IssueMessage(0,CONSOLE_AUTOCORRECT,_spritenum);
 			IssueMessage(0,AUTOCORRECTING,1,"num-defs",status.seensprites,length/4);
-			data.SetByteAt(1,uchar(status.seensprites=length/4));
+			data.SetByteAt(1,status.seensprites=length/4);
 		}
 		if(!status.seensprites)IssueMessage(WARNING1,NO_SETS,0x12);
 		if(CheckLength(length,2+4*status.seensprites,BAD_LENGTH,"byte 1","%2x",status.seensprites,2+4*status.seensprites))
